Include updates from Directory Chair.
Date: Sun, 13 Oct 2013 13:19:51 -0500
Subject: Re: directory
From: Kate Searcy <searcyk@gmail.com>
To: Chris Rorvick <chris@rorvick.com>

Chris,
We are definitely lucky to have a great person to work with on our
yearly projects! Thanks for the information; it was extremely helpful
and easy to understand.

I am attaching the updated spreadsheet. One update that we need to make
sure happens in the spelling of Kaui is really Kauai King Court. I
changed it on the spreadsheet. I wasn't sure if we needed to change it
any place else. (The map is correct).

As for the photo, do you have the one from last year? If not, I will
have one taken this afternoon edited and send to you ASAP!  What format
do you need me to put the photo in? I can send it to you in any, jpeg,
GIF, PDF, TIFF, etc. whatever is easiest! (That is something I know how
to do :) )
Kate