Automatically add continuation headers.
Use marks to add continuation headers automatically.  This will probably
need some further work at some point as this probably does not do the
right thing if the page break comes just before the next header.  But
that is not a problem now, so we can deal with it when it is.